Move players to spec after 3 minutes
I personally hate the afk kicker -.-
but the majority gets tripped up by afks and then they call
kickvotes which ban players longer than it would have to be.
Also the votes might not pass.
And if moderators go afk they can not even be voted.

Its sad for me that it takes away the utown city vibe
where you can afk at 2am towering a friend for 60 minutes.
That is also why i only roll it out to solofng and not to common
configs. Because on other servers there are no complaints about afks.